+++ src/geo/projection/vertical_perspective_camera_helper.ts
